added PaintJob solution

+++ PaintJob/PaintJob.java
@@ -1,6 +1,6 @@
 public class PaintJob {
     
-    public static int getBucketCount(double width, double height, double areaPerBucket, double extraBuckets){
+    public static int getBucketCount(double width, double height, double areaPerBucket, int extraBuckets){
 				
 		if(width<=0 || height<=0 || areaPerBucket<=0 || extraBuckets<0){
 			return -1;
